Refactors for public release

+++ parallel_pre_filter/src/constants.f90
@@ -1,13 +1,13 @@
-    module constants
+module constants
 
-      use precision_kinds
-      
-      implicit none
+  use precision_kinds
+  
+  implicit none
 
-      real(kind=dp) :: pi = 4.0_dp*datan(1.0_dp)
+  double precision :: pi = 4.0_dp*datan(1.0_dp)
 
-      real(kind=dp) :: mu = 1.32712440018e11 ! km^3/s^-2 
+  double precision :: mu = 1.32712440018e11 ! km^3/s^-2 
 
-      real(kind=dp) :: au = 149597870.7000 ! km
+  double precision :: au = 149597870.7000 ! km
 
-    end module constants
+end module constants
